+++ pkg/cmd/pipeline/logs.go
@@ -25,7 +25,7 @@ import (
 	"github.com/tektoncd/cli/pkg/options"
 	"github.com/tektoncd/cli/pkg/helper/pipeline"
 	phelper "github.com/tektoncd/cli/pkg/helper/pipeline"
-	prhelper "github.com/tektoncd/cli/pkg/helper/pipelinerun"
+	prhelper "github.com/tektoncd/cli/pkg/pipelinerun"
 	validate "github.com/tektoncd/cli/pkg/validate"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
