+++ pkg/cmd/pipelinerun/describe.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/tektoncd/cli/pkg/cli"
-	prdesc "github.com/tektoncd/cli/pkg/helper/pipelinerun/description"
+	prdesc "github.com/tektoncd/cli/pkg/pipelinerun/description"
 	validate "github.com/tektoncd/cli/pkg/validate"
 	"github.com/tektoncd/cli/pkg/printer"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
